:wrench: config file added
